Update #2
The engine now utilizes the library spdlog which allows for logging in the engine. There are two logger objects created, one for client and one for the engine.

The init function for the logging is currently in the application but it should be moved and initialized on the engine side.

Macros for calling the individual logger object along with the different log functions are created in Log.hpp which is inherited in the Hazel.hpp and then accessible though the EntryPoint.hpp.